refactor: rename e2e package to better follow the general convention. Delete tests not working yet
